Add multi-user support. Remove viewData and updateRecord

+++ lambda/delete-record.js
@@ -10,13 +10,26 @@ function standardHandler (context) {
   }
 }
 
-function deleteRecord (dynamo, params, context) {
+function deleteRecord (dynamo, params, context, userID) {
   if (!params.id) return context.fail('Missing param.id')
 
-  dynamo.deleteItem({
+  dynamo.getItem({
     'TableName': 'rcrd-records',
     'Key': {id: params.id}
-  }, standardHandler(context))
+  }, function (err, data) {
+    if (err) return context.fail(err)
+    var record = data.Item
+    if (!record) return context.fail('Record not found')
+
+    if (record.user_id !== userID) {
+      return context.fail('Insufficient privileges')
+    }
+
+    dynamo.deleteItem({
+      'TableName': 'rcrd-records',
+      'Key': {id: params.id}
+    }, standardHandler(context))
+  })
 }
 
 module.exports = deleteRecord
